backend/requirements_backend.txt
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==1.26.4
tensorflow-cpu
mlflow>=2.13.0
